Adaptado teste de criacao simples de Format e realizado algumas correcoes
