+++ tests/config.php
@@ -2,6 +2,4 @@
 
 require __DIR__.'/../vendor/autoload.php';
 
-define('CB_DIR_CB', __DIR__.'/../cb/');
-
 date_default_timezone_set('America/Fortaleza');
